all wallets fields as float64

+++ iperpetual/wallet.go
@@ -12,19 +12,19 @@ func (this WalletBalance) Do(client *Client) (map[string]Balance, bool) {
 }
 
 type Balance struct {
-	Equity           float32 `json:"equity"`
-	AvailableBalance float32 `json:"available_balance"`
-	UsedMargin       float32 `json:"used_margin"`
-	OrderMargin      float32 `json:"order_margin"`
-	PositionMargin   float32 `json:"position_margin"`
-	OccClosingFee    float32 `json:"occ_closing_fee"`
-	OccFundingFee    float32 `json:"occ_funding_fee"`
-	WalletBalance    float32 `json:"wallet_balance"`
-	RealisedPnl      float32 `json:"realised_pnl"`
-	UnrealisedPnl    float32 `json:"unrealised_pnl"`
-	CumRealisedPnl   float32 `json:"cum_realised_pnl"`
-	GivenCash        float32 `json:"given_cash"`
-	ServiceCash      float32 `json:"service_cash"`
+	Equity           float64 `json:"equity"`
+	AvailableBalance float64 `json:"available_balance"`
+	UsedMargin       float64 `json:"used_margin"`
+	OrderMargin      float64 `json:"order_margin"`
+	PositionMargin   float64 `json:"position_margin"`
+	OccClosingFee    float64 `json:"occ_closing_fee"`
+	OccFundingFee    float64 `json:"occ_funding_fee"`
+	WalletBalance    float64 `json:"wallet_balance"`
+	RealisedPnl      float64 `json:"realised_pnl"`
+	UnrealisedPnl    float64 `json:"unrealised_pnl"`
+	CumRealisedPnl   float64 `json:"cum_realised_pnl"`
+	GivenCash        float64 `json:"given_cash"`
+	ServiceCash      float64 `json:"service_cash"`
 }
 
 func (this *Client) WalletBalance(currency *string) (map[string]Balance, bool) {
